feat: use policy for post and comment

+++ app/Http/Controllers/CommentController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Post;
 use App\Models\Comment;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\StoreCommentRequest;
 use App\Http\Requests\UpdateCommentRequest;
 
@@ -24,6 +25,8 @@ class CommentController extends Controller
 
     public function update(UpdateCommentRequest $request, Comment $comment)
     {
+        Gate::authorize('update', $comment);
+
         $comment->update($request->validated());
 
         return back()->with('success', 'Comment updated.');
@@ -31,6 +34,8 @@ class CommentController extends Controller
     
     public function destroy(Comment $comment)
     {
+        Gate::authorize('delete', $comment);
+        
         $comment->delete();
         
         return back()->with('success', 'Comment deleted.');
